chore: finalize kest-core refactor, restore lab integration, and update benchmarks
- Export internal tools and models properly via kest.core public API
- Fix import failures in kest-lab (agent.py, app.py, gateway.py, tests)
- Update benchmark suite to invoke pure python pipeline natively
- Remove outdated Rust-based benchmark artifacts
- Append historical Rust benchmarks in BENCHMARK.md and ignore raw artifacts
