+++ libs/kest-core/python/examples/bench/.gitignore
@@ -1,0 +1,2 @@
+*.json
+*.txt
