+++ libs/kest-core/python/examples/bench/bench_throughput.py
@@ -27,14 +27,12 @@ from typing import Any
 # Allow running from the bench directory or from the project root.
 sys.path.insert(0, os.path.join(os.path.dirname(__file__), "../../"))
 
-from kest.core import KestEntry, get_backend, sign_entry
-from kest.core.identity.providers.local import LocalEd25519Provider
-from kest.core.models import Passport, PassportVerifier
+from kest.core import KestEntry, sign_entry, LocalEd25519Provider, Passport, PassportVerifier
 
 PROVIDER = LocalEd25519Provider()
 THREAD_COUNTS = [1, 2, 4, 8]
 WINDOW_SECS = 3
-BACKEND = get_backend()
+BACKEND = "python"
 
 # ---------------------------------------------------------------------------
 # Helpers
