--- libs/kest-core/python/examples/bench/decorator_throughput_python.json
@@ -1,9 +1,0 @@
-{
-  "backend": "python",
-  "decorator_throughput": {
-    "1": 8506.0,
-    "2": 6558.0,
-    "4": 8084.666666666667,
-    "8": 6529.333333333333
-  }
-}
