--- libs/kest-core/python/examples/bench/decorator_throughput_rust-v2.json
@@ -1,9 +1,0 @@
-{
-  "backend": "rust-v2",
-  "decorator_throughput": {
-    "1": 2026.6666666666667,
-    "2": 1913.0,
-    "4": 1937.6666666666667,
-    "8": 2031.3333333333333
-  }
-}
