--- libs/kest-core/python/examples/bench/decorator_throughput_rust.json
@@ -1,9 +1,0 @@
-{
-  "backend": "rust",
-  "decorator_throughput": {
-    "1": 4102.0,
-    "2": 4519.666666666667,
-    "4": 4197.0,
-    "8": 4355.333333333333
-  }
-}
