--- libs/kest-core/python/examples/bench/fastapi_rust_v2.txt
@@ -1,23 +1,0 @@
-
-# FastAPI Benchmark — backend=rust-v2
-  Window: 3s per test | Concurrency tested: [1, 2, 4, 8, 16, 32]
-
-## Scalability: Sync Endpoint (FastAPI ThreadPool)
- Threads |    ops/sec
--------------------------
-       1 |      580.0
-       2 |      538.0
-       4 |      476.7
-       8 |      337.7
-      16 |      255.3
-      32 |      259.0
-
-## Scalability: Async Endpoint (Single Event Loop)
-  Tasks  |    ops/sec
--------------------------
-       1 |      532.0
-       2 |      520.0
-       4 |      469.3
-       8 |      348.3
-      16 |      261.7
-      32 |      252.0
