--- libs/kest-core/python/examples/bench/summary.txt
@@ -1,13 +1,0 @@
-+----------------+---------+----------------------+-----------------------+
-| Benchmark      | rust    | python               | rust-v2               |
-+================+=========+======================+=======================+
-| sign_entry     | 21.5 us | not significant      | 22.3 us: 1.04x slower |
-+----------------+---------+----------------------+-----------------------+
-| chain_10       | 241 us  | 244 us: 1.01x slower | not significant       |
-+----------------+---------+----------------------+-----------------------+
-| Geometric mean | (ref)   | 1.01x slower         | 1.02x slower          |
-+----------------+---------+----------------------+-----------------------+
-
-Benchmark hidden because not significant (2): entry_create, canonical_json
-Ignored benchmarks (1) of examples/bench/python.json: chain_100
-Ignored benchmarks (1) of examples/bench/rust-v2.json: chain_100
