--- libs/kest-core/python/examples/bench/system_rust-v2.json
@@ -1,72 +1,0 @@
-{
-  "backend": "rust-v2",
-  "L1_baggage": {
-    "1": {
-      "pack_ms": 0.0027929698990192264,
-      "unpack_ms": 0.007557880016975105,
-      "raw_bytes": 829,
-      "packed_bytes": 829,
-      "tier": "inline \u2713"
-    },
-    "5": {
-      "pack_ms": 0.07765147005557083,
-      "unpack_ms": 0.06582088011782616,
-      "raw_bytes": 4145,
-      "packed_bytes": 936,
-      "tier": "compressed \u2713"
-    },
-    "10": {
-      "pack_ms": 0.033219729957636446,
-      "unpack_ms": 0.06548831996042281,
-      "raw_bytes": 8290,
-      "packed_bytes": 1120,
-      "tier": "compressed \u2713"
-    },
-    "25": {
-      "pack_ms": 0.05994258011924103,
-      "unpack_ms": 0.15578892998746596,
-      "raw_bytes": 20725,
-      "packed_bytes": 1492,
-      "tier": "compressed \u2713"
-    },
-    "50": {
-      "pack_ms": 0.09907105995807797,
-      "unpack_ms": 0.2978204400278628,
-      "raw_bytes": 41450,
-      "packed_bytes": 2052,
-      "tier": "compressed \u2713"
-    },
-    "claim_check_at_hop": null
-  },
-  "L2_decorator": {
-    "baseline_ms": 5.6110002333298326e-05,
-    "no_policy_ms": 0.39064519995008595,
-    "cedar_cold_ms": 1.4586510078515857,
-    "cedar_warm_ms": 0.46964918001322076,
-    "rego_cold_ms": 2.071953989798203,
-    "rego_warm_ms": 0.8366196000133641
-  },
-  "L3_policy": {
-    "cedar_ops_per_sec": 20497.333333333332,
-    "sign_ops_per_sec": 12401.666666666666
-  },
-  "L4_multihop": {
-    "3": {
-      "total_ms": 0.44989242014708,
-      "per_hop_ms": 0.14996414004902667
-    },
-    "5": {
-      "total_ms": 0.7934598799329251,
-      "per_hop_ms": 0.15869197598658502
-    },
-    "10": {
-      "total_ms": 1.9633854998392053,
-      "per_hop_ms": 0.19633854998392053
-    }
-  },
-  "L4_merge": {
-    "2": 0.011068219901062548,
-    "5": 0.02582482004072517,
-    "10": 0.04084264000994153
-  }
-}
