--- libs/kest-core/python/examples/bench/system_rust.json
@@ -1,72 +1,0 @@
-{
-  "backend": "rust",
-  "L1_baggage": {
-    "1": {
-      "pack_ms": 0.0028149900026619434,
-      "unpack_ms": 0.008408030116697773,
-      "raw_bytes": 829,
-      "packed_bytes": 829,
-      "tier": "inline \u2713"
-    },
-    "5": {
-      "pack_ms": 0.020526899897959083,
-      "unpack_ms": 0.036201449984218925,
-      "raw_bytes": 4145,
-      "packed_bytes": 940,
-      "tier": "compressed \u2713"
-    },
-    "10": {
-      "pack_ms": 0.03086782991886139,
-      "unpack_ms": 0.06656559999100864,
-      "raw_bytes": 8290,
-      "packed_bytes": 1108,
-      "tier": "compressed \u2713"
-    },
-    "25": {
-      "pack_ms": 0.05375972992624156,
-      "unpack_ms": 0.15440727991517633,
-      "raw_bytes": 20725,
-      "packed_bytes": 1376,
-      "tier": "compressed \u2713"
-    },
-    "50": {
-      "pack_ms": 0.09926311002345756,
-      "unpack_ms": 0.30601117003243417,
-      "raw_bytes": 41450,
-      "packed_bytes": 2056,
-      "tier": "compressed \u2713"
-    },
-    "claim_check_at_hop": null
-  },
-  "L2_decorator": {
-    "baseline_ms": 5.9250014601275325e-05,
-    "no_policy_ms": 0.18427085000439547,
-    "cedar_cold_ms": 1.3481250061886385,
-    "cedar_warm_ms": 0.32148104000953026,
-    "rego_cold_ms": 1.9835059938486665,
-    "rego_warm_ms": 0.19649894995382056
-  },
-  "L3_policy": {
-    "cedar_ops_per_sec": 20807.666666666668,
-    "sign_ops_per_sec": 13148.0
-  },
-  "L4_multihop": {
-    "3": {
-      "total_ms": 0.4293510000570677,
-      "per_hop_ms": 0.14311700001902258
-    },
-    "5": {
-      "total_ms": 0.8513284398941323,
-      "per_hop_ms": 0.17026568797882646
-    },
-    "10": {
-      "total_ms": 2.115416760207154,
-      "per_hop_ms": 0.21154167602071539
-    }
-  },
-  "L4_merge": {
-    "2": 0.02558646010584198,
-    "5": 0.0306923700554762,
-    "10": 0.05776667007012293
-  }
-}
