--- libs/kest-core/python/examples/bench/throughput_python.json
@@ -1,29 +1,0 @@
-{
-  "backend": "python",
-  "sign_scalability": {
-    "1": 16416.333333333332,
-    "2": 16657.333333333332,
-    "4": 16694.666666666668,
-    "8": 16814.333333333332
-  },
-  "verify_scalability": {
-    "1": 33831.333333333336,
-    "2": 33322.0,
-    "4": 32850.0,
-    "8": 35393.666666666664
-  },
-  "gil_contention": {
-    "baseline": 16444.0,
-    "contested": 13336.333333333334,
-    "degradation_pct": 18.898483742803858
-  },
-  "payload_sensitivity": {
-    "small": 17001.666666666668,
-    "medium": 13704.333333333334,
-    "large": 6993.666666666667
-  },
-  "memory": {
-    "sign_peak_bytes": 4502,
-    "verify_peak_bytes": 8944
-  }
-}
