--- libs/kest-core/python/examples/bench/throughput_rust-v2.json
@@ -1,29 +1,0 @@
-{
-  "backend": "rust-v2",
-  "sign_scalability": {
-    "1": 7840.666666666667,
-    "2": 14179.333333333334,
-    "4": 11747.0,
-    "8": 11330.0
-  },
-  "verify_scalability": {
-    "1": 33742.0,
-    "2": 33492.333333333336,
-    "4": 32793.666666666664,
-    "8": 34823.666666666664
-  },
-  "gil_contention": {
-    "baseline": 11804.666666666666,
-    "contested": 830.0,
-    "degradation_pct": 92.96888236290732
-  },
-  "payload_sensitivity": {
-    "small": 7831.333333333333,
-    "medium": 5389.0,
-    "large": 1859.0
-  },
-  "memory": {
-    "sign_peak_bytes": 3731,
-    "verify_peak_bytes": 9057
-  }
-}
