--- libs/kest-core/python/examples/bench/throughput_rust.json
@@ -1,29 +1,0 @@
-{
-  "backend": "rust",
-  "sign_scalability": {
-    "1": 7872.0,
-    "2": 14371.333333333334,
-    "4": 12301.666666666666,
-    "8": 11490.666666666666
-  },
-  "verify_scalability": {
-    "1": 34139.0,
-    "2": 34149.0,
-    "4": 33657.666666666664,
-    "8": 35656.333333333336
-  },
-  "gil_contention": {
-    "baseline": 11905.333333333334,
-    "contested": 796.6666666666666,
-    "degradation_pct": 93.30832120058238
-  },
-  "payload_sensitivity": {
-    "small": 7799.666666666667,
-    "medium": 5380.666666666667,
-    "large": 1857.0
-  },
-  "memory": {
-    "sign_peak_bytes": 3731,
-    "verify_peak_bytes": 8944
-  }
-}
