+++ libs/kest-core/python/src/kest/core/models/__init__.py
@@ -1,0 +1,11 @@
+from kest.core._core import KestEntry
+from kest.core.models.passport import (
+    Passport,
+    PolicyDeviation,
+)
+
+__all__ = [
+    "KestEntry",
+    "Passport",
+    "PolicyDeviation",
+]
