+++ showcase/kest-lab/gateway.py
@@ -41,9 +41,10 @@ from kest.core import (
     LocalEd25519Provider,
     SPIREProvider,
     CedarLocalEngine,
+    KestMiddleware,
+    KestIdentityMiddleware,
 )
-from kest.core.ext import KestMiddleware, KestIdentityMiddleware
-import kest.core.decorators
+import kest.core.framework.decorators as kest_decorators
 
 # ---------------------------------------------------------------------------
 # OpenTelemetry
@@ -55,7 +56,7 @@ otlp_endpoint = os.getenv("OTEL_EXPORTER_OTLP_ENDPOINT", "http://otel-collector:
 exporter = OTLPSpanExporter(endpoint=f"{otlp_endpoint}/v1/traces")
 provider.add_span_processor(BatchSpanProcessor(exporter))
 trace.set_tracer_provider(provider)
-kest.core.decorators.tracer = trace.get_tracer("kest.core")
+kest_decorators.tracer = trace.get_tracer("kest.core")
 
 # ---------------------------------------------------------------------------
 # Identity Provider
